+++ forms-flow-api/src/formsflow_api/resources/application.py
@@ -181,6 +181,7 @@ class ApplicationResourceById(Resource):
     @staticmethod
     @auth.require
     @profiletime
+    @auth.has_one_of_roles(["camunda-client"])
     def delete(application_id):
         """Delete application by id."""
         try:
