+++ forms-flow-bpm/src/main/java/org/camunda/bpm/extension/hooks/listeners/DeleteSubmissionListener.java
@@ -1,6 +1,15 @@
 package org.camunda.bpm.extension.hooks.listeners;
 
 import org.camunda.bpm.engine.delegate.*;
+import org.camunda.bpm.extension.commons.connector.HTTPServiceInvoker;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.camunda.bpm.extension.hooks.exceptions.ApplicationServiceException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.JsonNode;
+
+
 import org.camunda.bpm.extension.hooks.services.FormSubmissionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -14,10 +23,14 @@ public class DeleteSubmissionListener extends BaseListener implements TaskListen
     @Autowired
     private FormSubmissionService formSubmissionService;
 
+    @Autowired
+    private HTTPServiceInvoker httpServiceInvoker;
+
     @Override
     public void notify(DelegateExecution execution) {
         try {
             deleteSubmission(execution);
+            deleteApplication(execution);
         } catch (IOException e) {
             handleException(execution, ExceptionSource.EXECUTION, e);
         }
@@ -27,6 +40,7 @@ public class DeleteSubmissionListener extends BaseListener implements TaskListen
     public void notify(DelegateTask delegateTask) {
         try {
             deleteSubmission(delegateTask.getExecution());
+            deleteApplication(delegateTask.getExecution());
         } catch (IOException e) {
             handleException(delegateTask.getExecution(), ExceptionSource.TASK, e);
         }
@@ -41,5 +55,23 @@ public class DeleteSubmissionListener extends BaseListener implements TaskListen
             System.out.println("Error deleting submission: " + e.getMessage());
             e.printStackTrace();
         }    
+    }
+
+    private String getApplicationUrl(DelegateExecution execution){
+        return httpServiceInvoker.getProperties().getProperty("api.url")+"/application/"+execution.getVariable("applicationId");
+    }
+
+    private void deleteApplication(DelegateExecution execution) throws IOException {
+        System.out.println("DeleteApplication started");
+        String applicationUrl = getApplicationUrl(execution);
+        ObjectMapper objectMapper = new ObjectMapper();
+        ResponseEntity<String> response = httpServiceInvoker.execute(applicationUrl, HttpMethod.DELETE, null);
+        if (response.getStatusCode().value() == HttpStatus.OK.value()) {
+            System.out.println("Application was deleted successfully: " + applicationUrl);
+            JsonNode jsonNode = objectMapper.readTree(response.getBody());
+        } else {
+            throw new ApplicationServiceException("Unable to delete application for: " + applicationUrl + ". Message Body: " +
+                    response.getBody());
+        }
     }
 }
